Rework exercise command
Internally, \exercise does not use \section anymore, but relies completely on its own type of section, defined by \@exercisesection. Like before, it does use the same counter (more specifically, an alias of it, so \autoref behaves correctly). While this should be a more elegant solution in general, the main reason for doing this was the following. Previously, since \refstepcounter was executed *after* \section (which was really necessary, as the 'section' counter set by \section would mess up the name for \autoref), PDF links jumped to the wrong position in the document. Now, the counter can be incremented before the exercise section is printed, fixing these misaligned jumps.